Move zipkin v2 trace conversion code into translator/trace/zipkin use same tag parsing logic for both zipkin v1 and v2
