+++ translator/trace/zipkin/status_code_test.go
@@ -93,3 +93,123 @@ func TestAttribToStatusCode(t *testing.T) {
 		})
 	}
 }
+
+func TestStatusCodeMapperCases(t *testing.T) {
+	tests := []struct {
+		name       string
+		expected   *tracepb.Status
+		attributes map[string]string
+	}{
+		{
+			name:     "no relevant attributes",
+			expected: nil,
+			attributes: map[string]string{
+				"not.relevant": "data",
+			},
+		},
+
+		{
+			name:     "http: 500",
+			expected: &tracepb.Status{Code: 13},
+			attributes: map[string]string{
+				"http.status_code": "500",
+			},
+		},
+
+		{
+			name:     "http: 500",
+			expected: &tracepb.Status{Code: 13, Message: "a message"},
+			attributes: map[string]string{
+				"http.status_code":    "500",
+				"http.status_message": "a message",
+			},
+		},
+
+		{
+			name:     "http: 500, with error attribute",
+			expected: &tracepb.Status{Code: 13},
+			attributes: map[string]string{
+				"http.status_code": "500",
+				"error":            "an error occurred",
+			},
+		},
+
+		{
+			name:     "oc: internal",
+			expected: &tracepb.Status{Code: 13, Message: "a description"},
+			attributes: map[string]string{
+				"census.status_code":        "13",
+				"census.status_description": "a description",
+			},
+		},
+
+		{
+			name:     "oc: description and error",
+			expected: &tracepb.Status{Code: 13, Message: "a description"},
+			attributes: map[string]string{
+				"opencensus.status_description": "a description",
+				"error":                         "INTERNAL",
+			},
+		},
+
+		{
+			name:     "oc: error only",
+			expected: &tracepb.Status{Code: 13, Message: ""},
+			attributes: map[string]string{
+				"error": "INTERNAL",
+			},
+		},
+
+		{
+			name:     "oc: description only, use unknown code",
+			expected: &tracepb.Status{Code: 2, Message: "a description"},
+			attributes: map[string]string{
+				"opencensus.status_description": "a description",
+			},
+		},
+
+		{
+			name:     "oc: priority over http",
+			expected: &tracepb.Status{Code: 4, Message: "deadline expired"},
+			attributes: map[string]string{
+				"census.status_description": "deadline expired",
+				"census.status_code":        "4",
+
+				"http.status_message": "a description",
+				"http.status_code":    "500",
+			},
+		},
+
+		{
+			name:     "error only",
+			expected: nil,
+			attributes: map[string]string{
+				"error": "a description",
+			},
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			attributes := attributesFromMap(test.attributes)
+
+			sMapper := &statusMapper{}
+			for k, v := range attributes {
+				sMapper.fromAttribute(k, v)
+			}
+
+			got := sMapper.ocStatus()
+			assert.EqualValues(t, test.expected, got)
+		})
+	}
+}
+
+func attributesFromMap(mapValues map[string]string) map[string]*tracepb.AttributeValue {
+	res := map[string]*tracepb.AttributeValue{}
+
+	for k, v := range mapValues {
+		pbAttrib := parseAnnotationValue(v)
+		res[k] = pbAttrib
+	}
+	return res
+}
