+++ translator/trace/zipkin/testdata/zipkin_v2_single.json
@@ -1,0 +1,38 @@
+[
+  {
+    "traceId": "4d1e00c0db9010db86154a4ba6e91385",
+    "parentId": "86154a4ba6e91385",
+    "id": "4d1e00c0db9010db",
+    "kind": "CLIENT",
+    "name": "get",
+    "timestamp": 1472470996199000,
+    "duration": 207000,
+    "localEndpoint": {
+      "serviceName": "frontend",
+      "ipv6": "7::0.128.128.127"
+    },
+    "remoteEndpoint": {
+      "serviceName": "backend",
+      "ipv4": "192.168.99.101",
+      "port": 9000
+    },
+    "annotations": [
+      {
+        "timestamp": 1472470996238000,
+        "value": "foo"
+      },
+      {
+        "timestamp": 1472470996403000,
+        "value": "bar"
+      }
+    ],
+    "tags": {
+      "http.path": "/api",
+      "http.status_code": "500",
+      "cache_hit": "true",
+      "ping_count": "25",
+      "timeout": "12.3",
+      "clnt/finagle.version": "6.45.0"
+    }
+  }
+]
